feat: extract duration_format module to separate file
